+++ CHANGELOG.md
@@ -149,6 +149,7 @@ All notable changes to this project will be documented in this file.
 - Axis should never go into infinite loop (#758)
 - Exception in BarSeriesBase (#790)
 - Vertical Axes Title Font Bug (#474)
+- Support string[] as ItemsSource in CategoryAxis #825
 
 ## [2014.1.546] - 2014-10-22
 ### Added